refactor: move AppData to own module
