+++ backend/src/config/mod.rs
@@ -3,4 +3,5 @@
 pub mod api_doc;
 pub mod app;
 pub mod auth;
+pub mod data;
 pub mod routes;
