+++ backend/src/controller/plantings.rs
@@ -14,13 +14,13 @@ use actix_web::{
 };
 
 use crate::config::auth::user_info::UserInfo;
+use crate::config::data::AppDataInner;
 use crate::model::dto::actions::{
     CreatePlantActionDto, DeletePlantActionDto, MovePlantActionDto, TransformPlantActionDto,
 };
 use crate::model::dto::{
     NewPlantingDto, PlantLayerObjectDto, PlantingSearchParameters, UpdatePlantingDto,
 };
-use crate::AppDataInner;
 
 /// FIXME: REMOVE THIS HACK ❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗❗
 static PLANTINGS: RwLock<Vec<PlantLayerObjectDto>> = RwLock::new(vec![]);
